# contracts/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plant Doctor RAG API
  version: 0.2.0
  description: |
    식물 주치의 AI MVP용 프론트-백엔드 공유 API 계약 문서입니다.
    
    [업데이트 히스토리]
    * v0.2.0: Supabase pgvector 클라우드 DB 연동 및 RAG 상담 이력 메시지 스키마 정합 반영 완료.
    * v0.1.0: RAG 상담 및 이력 조회 API 엔드포인트 초안 수립.
servers:
  - url: http://localhost:8000
    description: Local backend
security:
  - bearerAuth: []
paths:
...
  /api/v1/uploads/signed-url:
    post:
      summary: Get signed upload URL for plant photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadSignedUrlRequest"
      responses:
        "200":
          description: Signed upload URL and storage path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadSignedUrlResponse"
  /api/v1/uploads/plant-photo:
    post:
      summary: Upload plant photo through backend and register metadata
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PlantPhotoUploadRequest"
      responses:
        "201":
          description: Uploaded and registered photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantPhoto"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    PlantPhotoCreate:
      type: object
      required: [storagePath]
      properties:
        storagePath:
          type: string
        capturedAt:
          type: string
          format: date-time
        note:
          type: string
    PlantPhoto:
      allOf:
        - $ref: "#/components/schemas/PlantPhotoCreate"
        - type: object
          required: [id, plantId, createdAt]
          properties:
            id:
              type: string
              format: uuid
            plantId:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
...
    UploadSignedUrlRequest:
      type: object
      required: [fileName]
      properties:
        fileName:
          type: string
          example: monstera.jpg
        mimeType:
          type: string
          example: image/jpeg
    UploadSignedUrlResponse:
      type: object
      required: [signedUrl, storagePath]
      properties:
        signedUrl:
          type: string
          example: https://xyz.supabase.co/storage/v1/object/upload/sign/plant-photos/...
        storagePath:
          type: string
          example: users/d3b07384-d113-49c3-a558-1ec114a84d41/plants/abcd-efgh-1234_monstera.jpg
    PlantPhotoUploadRequest:
      type: object
      required: [plantId, file]
      properties:
        plantId:
          type: string
          format: uuid
        note:
          type: string
        capturedAt:
          type: string
          format: date-time
        file:
          type: string
          format: binary
...